Advanced extensibility support ==============================
One problem with our current pre.js model is that it can lead to maccaroni code
- just like spaghetti code, it can get messy
- but it's made of pipelines
- and you try to squeeze a lot of functionality through a single extension point

With this model, we can open multiple extension points along the pipeline, so that
a developer can inject custom functions before or after a particular pipeline step,
leading to better-stuctured penne code.

The attaching works through the `Pipeline.attach.before` and `Pipeline.attach.after`
functions which take each a pair of a function (like your pre function) and the name
of a pipeline step.

To make things easier, I plan to extend the wrapper in `helix-cli`, so that it carries
over `module.exports.before` and `module.exports.after` in a `pre.js`, just like
`module.exports.pre`. This would enable easy declarative extensibility without having
to override the entire pipeline.